requirements.txt
python-telegram-bot==13.*
python-dotenv==1.*
google-cloud-dialogflow==2.21.*
vk_api==11.9.*